split view tests in to own files
